Move Redux actions into reducer files
