--- assets/js/state/actions/host.js
@@ -1,1 +1,0 @@
-export const HOST_DEREGISTERED = 'HOST_DEREGISTERED';
