--- assets/js/state/actions/sapSystem.js
@@ -1,7 +1,0 @@
-export const SAP_SYSTEM_REGISTERED = 'SAP_SYSTEM_REGISTERED';
-export const SAP_SYSTEM_HEALTH_CHANGED = 'SAP_SYSTEM_HEALTH_CHANGED';
-export const APPLICATION_INSTANCE_REGISTERED =
-  'APPLICATION_INSTANCE_REGISTERED';
-export const APPLICATION_INSTANCE_HEALTH_CHANGED =
-  'APPLICATION_INSTANCE_HEALTH_CHANGED';
-export const SAP_SYSTEM_DEREGISTERED = 'SAP_SYSTEM_DEREGISTERED';
